Move from yacl to zen config

+++ src/main/kotlin/meowing/zen/feats/general/blockoverlay.kt
@@ -23,7 +23,7 @@ object blockoverlay : Feature("blockoverlay") {
             if (blockShape.isEmpty) return@register
 
             val camPos = camera.pos
-            RenderSystem.lineWidth(Zen.config.blockoverlaywidth)
+            RenderSystem.lineWidth(Zen.config.blockoverlaywidth.toFloat())
             event.cancel()
 
             VertexRendering.drawOutline(
